requirements.txt
fastapi==0.128.2
uvicorn==0.29.0
pydantic==2.12.5
pypdf===6.6.2
python-docx===1.2.0
celery===5.6.2
redis===7.2.0